intrusion-management-api/requirements.txt
fastapi==0.85.1
mariadb==1.0.11
uvicorn==0.18.3
SQLAlchemy==1.4.41
pydantic==1.10.2
boto3==1.20.10
python-multipart==0.0.5
python-dotenv==0.19.1
kombu==5.2.4
alembic==1.8.1
pytest
requests
pytest-cov
urllib3
tox
google-api-python-client
google-auth-httplib2
google-auth-oauthlib